+++ test/integration/decoder/decoder_test.h
@@ -1,0 +1,8 @@
+/* Decoder Integration Test Header */
+
+#ifndef DECODER_TEST_H
+#define DECODER_TEST_H
+
+extern void test_decoder(void);
+
+#endif